Refactored ImageInformation into ImageInformation_v2_1_1
