+++ cineast-core/src/test/java/org/vitrivr/cineast/core/iiif/imageapi/ImageInformationV211RequestTest.java
